+++ src/datarobot_genai/drmcp/__init__.py
@@ -28,15 +28,15 @@ from datarobot_genai.drmcp.test_utils.mcp_utils_integration import integration_t
 from datarobot_genai.drmcp.test_utils.tool_base_ete import ETETestExpectations
 from datarobot_genai.drmcp.test_utils.tool_base_ete import ToolBaseE2E
 from datarobot_genai.drmcp.test_utils.tool_base_ete import ToolCallTestExpectations
-from datarobot_genai.drtools.core.clients.datarobot import request_user_dr_sdk
+from datarobot_genai.drmcputils.clients.datarobot import request_user_dr_sdk
+from datarobot_genai.drmcputils.constants import RUNTIME_PARAM_ENV_VAR_NAME_PREFIX
+from datarobot_genai.drmcputils.credentials import ToolsAuthCredentials
+from datarobot_genai.drmcputils.credentials import get_credentials
 from datarobot_genai.drtools.core.config_utils import (
     extract_datarobot_credential_runtime_param_payload,
 )
 from datarobot_genai.drtools.core.config_utils import extract_datarobot_dict_runtime_param_payload
 from datarobot_genai.drtools.core.config_utils import extract_datarobot_runtime_param_payload
-from datarobot_genai.drtools.core.constants import RUNTIME_PARAM_ENV_VAR_NAME_PREFIX
-from datarobot_genai.drtools.core.credentials import ToolsAuthCredentials
-from datarobot_genai.drtools.core.credentials import get_credentials
 
 from .core.config import MCPServerConfig
 from .core.config import get_config